fixed rectangle rendering

+++ OpenXaml/include/OpenXaml/Environment/Window.h
@@ -9,8 +9,8 @@ namespace OpenXaml
 		public:
 			float xScale = 0.0f;
 			float yScale = 0.0f;
-			unsigned int width = 0;
-			unsigned int height = 0;
+			float width = 0;
+			float height = 0;
 		};
 
 		extern Window* window;
